Add support for FATE data encoding
- split TypeResolver from AciTypeResolver
- add public API TypeResolver to support working with type information
- add encode and decode support with type information to ContractByteArrayEncoder
