+++ src/main.js
@@ -4,6 +4,7 @@ const BytecodeContractCallEncoder = require('./api/BytecodeContractCallEncoder')
 const ContractByteArrayEncoder = require('./api/ContractByteArrayEncoder')
 const FateApiEncoder = require('./api/FateApiEncoder')
 const ContractEncoder = require('./api/ContractEncoder')
+const TypeResolver = require('./api/TypeResolver')
 
 module.exports = {
     Encoder,
@@ -12,4 +13,5 @@ module.exports = {
     ContractByteArrayEncoder,
     FateApiEncoder,
     ContractEncoder,
+    TypeResolver,
 }
